Removed setting the body to display:none
I think it is better to remove this feature as it can do more harm then good.
For example if you try to get the height of an element with jQuery('#mydiv').height() the method returns 0 because the body is not displayed yet. 

Also not showing the body when loading can give the user the false idea that the page isn't loading at all.